Refactor FlixelBaseVideo into FlixelVideo, fix video volume and auto-pause
FlixelBaseVideo is dissolved; its content moves into a new FlixelVideo
abstract class, replacing the now-redundant FlixelVideo interface. All
backends extend FlixelVideo directly.

A LifecycleListener registered in the FlixelVideo constructor pauses
each video when the OS suspends the application (respecting autoPause)
and resumes it when the application returns to the foreground, matching
the behavior FlixelGame applies to sounds.

Volume on Android and web is now re-asserted unconditionally every frame
inside updateMedia(), matching the existing desktop VLC approach. This
prevents audio focus changes (Android) and browser audio context
initialization (web) from silently resetting a video back to full volume.
